% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectors.R
\name{license_list}
\alias{license_list}
\title{Get all OSI-listed licenses}
\usage{
license_list(...)
}
\arguments{
\item{...}{arguments to pass to httr's GET.}
}
\value{
a list of licenses and their metadata (themselves in the form of
lists) which can be provided to other functions such as \code{\link{license_text}}
}
\description{
\code{license_list} retrieves a list of
all the metadata associated with every license that the Open Source Initiative
has coverage of.
}
\examples{
# Get licenses
list_of_all_licenses <- license_list()

}

